+++ package.json
@@ -28,34 +28,35 @@
     "url": "https://github.com/jdi-testing/jdn"
   },
   "dependencies": {
-    "antd": "^4.2.4",
+    "@ant-design/icons": "^4.6.2",
+    "antd": "^4.14.0",
     "classnames": "^2.2.6",
-    "jszip": "^3.4.0",
-    "mobx": "^5.15.4",
-    "mobx-react": "^6.2.2",
+    "jszip": "^3.6.0",
+    "mobx": "^6.1.8",
+    "mobx-react": "^7.1.0",
     "mobx-react-devtools": "^6.1.1",
-    "parcel-bundler": "^1.12.4",
+    "parcel-bundler": "^1.12.5",
     "primer": "^11.0.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
     "react-file-reader": "^1.1.4",
-    "react-jss": "^10.1.1",
+    "react-jss": "^10.6.0",
     "react-loading": "^2.0.3",
-    "react-responsive": "^8.0.3",
-    "react-select": "^3.1.0",
-    "simple-crypto-js": "^2.2.0"
+    "react-responsive": "^8.2.0",
+    "react-select": "^4.3.0",
+    "simple-crypto-js": "^3.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.9.0",
-    "@babel/plugin-proposal-class-properties": "^7.8.3",
-    "@babel/plugin-proposal-decorators": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.9.0",
-    "@babel/preset-env": "^7.9.5",
-    "@babel/preset-react": "^7.9.4",
-    "@babel/runtime": "^7.9.2",
-    "@svgr/webpack": "^5.4.0",
-    "babel-jest": "^25.4.0",
-    "babel-loader": "^8.1.0",
+    "@babel/core": "^7.13.10",
+    "@babel/plugin-proposal-class-properties": "^7.13.0",
+    "@babel/plugin-proposal-decorators": "^7.13.5",
+    "@babel/plugin-transform-runtime": "^7.13.10",
+    "@babel/preset-env": "^7.13.10",
+    "@babel/preset-react": "^7.12.13",
+    "@babel/runtime": "^7.13.10",
+    "@svgr/webpack": "^5.5.0",
+    "babel-jest": "^26.6.3",
+    "babel-loader": "^8.2.2",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.5",
     "babel-plugin-transform-object-rest-spread": "^7.0.0-beta.3",
@@ -64,27 +65,27 @@
     "babel-preset-mobx": "^2.0.0",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-1": "^6.24.1",
-    "css-loader": "^3.5.3",
-    "eslint": "^7.0.0",
+    "css-loader": "^5.1.3",
+    "eslint": "^7.22.0",
     "eslint-config-google": "^0.14.0",
-    "eslint-plugin-react": "^7.20.0",
+    "eslint-plugin-react": "^7.22.0",
     "extract-loader": "^5.1.0",
     "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "^6.0.0",
-    "file-saver": "^2.0.2",
-    "html-webpack-plugin": "^4.3.0",
-    "jest": "^25.4.0",
+    "file-loader": "^6.2.0",
+    "file-saver": "^2.0.5",
+    "html-webpack-plugin": "^5.3.1",
+    "jest": "^26.6.3",
     "json-loader": "^0.5.7",
-    "less-loader": "^6.1.0",
+    "less-loader": "^8.0.0",
     "manifest-loader": "0.0.1",
-    "mini-css-extract-plugin": "^0.9.0",
-    "prettier": "2.0.5",
-    "react-test-renderer": "^16.13.1",
+    "mini-css-extract-plugin": "^1.3.9",
+    "prettier": "2.2.1",
+    "react-test-renderer": "^17.0.1",
     "recursive-readdir": "^2.2.2",
-    "resolve-url-loader": "^3.1.1",
-    "style-loader": "^1.2.1",
-    "webpack": "^4.42.1",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.10.3"
+    "resolve-url-loader": "^3.1.2",
+    "style-loader": "^2.0.0",
+    "webpack": "^5.27.1",
+    "webpack-cli": "^4.5.0",
+    "webpack-dev-server": "^3.11.2"
   }
 }
