+++ src/css/main.less
@@ -18,8 +18,85 @@ button svg {
 .ant-btn {
   border: solid 1px #777777;
   color: #464547;
+  &.ant-btn-link{
+    border: none;
+  }
+}
+
+.treeCounter {
+  padding: 0px 4px 0px 4px;
+  font-family: SourceSansPro;
+  font-size: 13px;
+  font-weight: normal;
+  font-stretch: normal;
+  font-style: normal;
+  line-height: normal;
+  letter-spacing: normal;
+  text-align: center;
+  color: #222222;  border-radius: 8px;
+  border: solid 1px #cbcbcb;
+  background-color: #ffffff;
+}
+
+.field{
+  font-family: SourceSansPro;
+  font-size: 11px;
+  font-weight: normal;
+  font-stretch: normal;
+  font-style: normal;
+  line-height: normal;
+  letter-spacing: normal;
+  color: #777777;
+  text-transform: capitalize;
+}
+
+.inputsRow{
+  display: inline-block;
+  & > div {
+    float: left;
+  }
+}
+
+.db{
+  display: block;
+}
+
+.posr{position: relative}
+
+.add {
+  position: absolute;
+  right: 0px;
+  top: -10px;
+
+}
+
+.ant-input {
+  color: #008ace
+}
+
+.list {
+  padding-left: 0px;
+  list-style-type: none;
+  display: block;
+
+  .list {
+    padding-left: 36px;
+  }
+}
+
+.link {
+  cursor: pointer;
+}
+
+.ant-btn-primary {
+  color: #fff;
+  border: none;
 }
 
 .ant-btn-sm {
   font-size: 12px;
+}
+
+.ant-select-selector {
+  border: 1px solid #d8d8d8 !important;
 }
