+++ src/js/blocks/generate/GenerateSettings.jsx
@@ -3,8 +3,8 @@ import injectSheet from "react-jss";
 import ReactFileReader from "react-file-reader";
 import PropTypes from "prop-types";
 import { inject, observer } from "mobx-react";
-import Button from "../../../components/Button/Button";
-import { exportIcon, importIcon, back } from "../../../../icons/index";
+import { exportIcon, importIcon, back } from "../../../icons";
+import { Button } from "antd";
 
 const styles = {
   generateStyle: {
@@ -24,8 +24,6 @@ const styles = {
 class GenerateSettings extends React.Component {
   handleSettings = () => {
     const { mainModel } = this.props;
-
-    mainModel.setRightPart("GeneralSettingsWrapper");
   };
 
   handleExportTemplate = () => {
@@ -40,28 +38,11 @@ class GenerateSettings extends React.Component {
     mainModel.generateBlockModel.clearGeneration();
     mainModel.settingsModel.importNewTemplate(file, mainModel);
   };
-
-  handleBack = () => {
-    const { mainModel } = this.props;
-
-    mainModel.setLeftPart("GenerateBlockWrapper");
-    if (mainModel.generateBlockModel.pages.length) {
-      mainModel.setRightPart("GenerateResultsWrapper");
-    } else {
-      mainModel.setRightPart();
-    }
-  };
-
   render() {
     const { classes } = this.props;
     return (
       <div>
         <div className={`${classes.generateStyle} BtnGroup`}>
-          <Button
-            className="BtnGroup-item"
-            icon={back}
-            onclick={this.handleBack}
-          />
           <Button
             className="BtnGroup-item"
             label={"Settings"}
