+++ src/js/models/MainModel.jsx
@@ -1,13 +1,7 @@
 import { observable, action } from "mobx";
 import GenerateBlockModel from "./GenerateBlockModel";
 import RulesBlockModel from "./RulesBlockModel";
-import GenerateBlockWrapper from "../blocks/generate/left/GenerateBlock";
-import GenerateSettingsWrapper from "../blocks/generate/left/GenerateSettings";
-import GenerateResultsWrapper from "../blocks/generate/right/GenerateResults";
-import GeneralSettingsWrapper from "../blocks/generate/right/GeneralSettings";
-import RulesBlockWrapper from "../blocks/rules/left/RulesBlock";
-import ListOfSearchAttributesWrapper from "../blocks/rules/right/ListOfSearchAttributes";
-import RuleForElementWrapper from "../blocks/rules/right/RuleForElement";
+import GenerateBlockWrapper from "../blocks/generate/GenerateBlock";
 import ConversionToCodeModel from "./ConversionToCodeModel";
 import SettingsModel from "./SettingsModel";
 
@@ -39,20 +33,6 @@ export default class MainModel {
   }
 
   @action
-  setRightPart(currentRightPart, currentRule, ruleSet) {
-    this.currentRightPart = currentRightPart;
-    // ruleSet ? this.ruleBlockModel.setCurrentRuleSet(ruleSet) : this.ruleBlockModel.setCurrentRuleSet('');
-    // currentRule ? this.ruleBlockModel.setCurrentRuleName(currentRule) : this.ruleBlockModel.setCurrentRuleName('');
-  }
-
-  @action
-  setLeftPart(currentLeftPart) {
-    this.currentLeftPart = currentLeftPart;
-    // ruleSet ? this.ruleBlockModel.setCurrentRuleSet(ruleSet) : this.ruleBlockModel.setCurrentRuleSet('');
-    // currentRule ? this.ruleBlockModel.setCurrentRuleName(currentRule) : this.ruleBlockModel.setCurrentRuleName('');
-  }
-
-  @action
   clearLog() {
     this.applicationLog = [];
   }
@@ -66,9 +46,4 @@ export default class MainModel {
   triggerShowLog() {
     this.showLog = !this.showLog;
   }
-
-  // @action
-  // setPageId (id) {
-  // 	this.currentPageId = id;
-  // }
 }
